requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Async Tasks
celery==5.3.4
django-celery-beat==2.5.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# Payment Integration
yookassa==2.4.0

# Email/SMS
django-anymail==10.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.18.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
